BookStore/Application/GenreOperations/Commands/DeleteGenre/DeleteGenreCommand.cs
using BookStore.DBOperations;
using System;
using System.Linq;

namespace BookStore.Application.GenreOperations.Commands.DeleteGenre
{
    public class DeleteGenreCommand
    {
        private readonly IBookStoreDbContext _dbContext;
        public int GenreId { get; set; }

        public DeleteGenreCommand(IBookStoreDbContext dbContext)
        {
            _dbContext = dbContext;
        }

        public void Handle()
        {
            var genre = _dbContext.Genres.SingleOrDefault(x => x.Id == GenreId);
            if (genre is null)
                throw new InvalidOperationException("Silinecek kitap türü bulunamadı!");


            _dbContext.Genres.Remove(genre);
            _dbContext.SaveChanges();
        }
    }
}
